change folder architecture
