+++ .github/CODEOWNERS
@@ -1,0 +1,2 @@
+# Tous les fichiers doivent être approuvés par @C-Alexis4414 et @accardigianni
+* @C-Alexis4414 @accardigianni
