--- ltef/.github/CODEOWNERS
@@ -1,2 +1,0 @@
-# Tous les fichiers doivent être approuvés par @C-Alexis4414 et @accardigianni
-* @C-Alexis4414 @accardigianni
